Atas de reuniões dos Comitês de Bacia Hidrográfica no Estado de São Paulo

Tabela contendo informações sobre reuniões dos Comitês de Bacia Hidrográfica no Estado de São Paulo e link para as atas de reunião.

atas_comites

Format

Uma tibble, contendo:

data_coleta_dados

Data em que os dados foram coletados, no formato ano-mês-dia (yyyy-mm-dd)

site_coleta

Link do site onde os dados foram coletados

comite

Nome do comitê de bacia

n_ugrhi

Número que representa a Unidade de Gerenciamento de Recursos Hídricos à qual o Comitê está relacionado

nome_reuniao

Nome da reunião, segundo informado pelo Comitê de Bacia

data_reuniao

Data da reunião, segundo informado pelo Comitê de Bacia, no formato ano-mês-dia (yyyy-mm-dd)

data_postagem

Data da postagem da ata de reunião, segundo informado pelo Comitê de Bacia, no formato ano-mês-dia (yyyy-mm-dd)

numero_link

Em alguns casos, há mais de um link para o arquivo da ata por reunião. Essa coluna contém a ordem dos arquivos.

url_link

URL, ou endereço online, do arquivo referente à ata da reunião.

Source

SigRH - http://www.sigrh.sp.gov.br/

Examples

atas_comites %>% dplyr::arrange(desc(data_postagem)) %>% head(10)

#> # A tibble: 10 x 9
#>    data_coleta_dados site_coleta   comite  n_ugrhi nome_reuniao     data_reuniao
#>    <date>            <glue>        <chr>     <dbl> <chr>            <date>      
#>  1 2020-10-21        http://www.s… Piraci…       5 Ata da 10ª Reun… 2020-06-30  
#>  2 2020-10-21        http://www.s… Piraci…       5 Ata da 10ª Reun… 2020-06-30  
#>  3 2020-10-21        http://www.s… Litora…       3 Ata da 3ª Reuni… 2019-12-13  
#>  4 2020-10-21        http://www.s… Baixo …      12 ATA DA 47º REUN… 2019-12-17  
#>  5 2020-10-21        http://www.s… Piraci…       5 Ata da 9ª Reuni… 2019-12-13  
#>  6 2020-10-21        http://www.s… Piraci…       5 Ata da 9ª Reuni… 2019-12-13  
#>  7 2020-10-21        http://www.s… Médio …      17 Ata da 41ª Reun… 2019-11-04  
#>  8 2020-10-21        http://www.s… Aquape…      21 Ata da 38ª Reun… 2019-11-04  
#>  9 2020-10-21        http://www.s… Alto T…       6 Resumo Executiv… 2020-02-20  
#> 10 2020-10-21        http://www.s… Tietê-…      16 Ata da 1ª Reuni… 2020-07-17  
#> # … with 3 more variables: data_postagem <date>, numero_link <chr>,
#> #   url_link <chr>