Obter tabelas sobre representantes dos Comitês de Bacia

Função para obter tabelas sobre representantes dos Comitês de Bacia no Estado de São Paulo

obter_tabela_representantes_comites(
  sigla_do_comite = NULL,
  online = TRUE,
  path_arquivo = NULL
)

Arguments

sigla_do_comite
Texto referente à sigla do comitê. É possível verificar na base: comites_sp.

online
Caso seja TRUE, a tabela será obtida consultando a página web. Caso seja FALSE, a tabela será obtida usando um caminho, que deve ser informado no argumento path_arquivo. O valor padrão é TRUE.

path_arquivo
Caminho para o arquivo .html que será lido. Isso só deve ser usado com o argumento online sendo FALSE. O caminho para o arquivo deve ser o gerado pela função download_html.

Value

Uma tibble. Uma base com dados coletados para todos os comitês está disponível em representantes_comites.

Examples

obter_tabela_representantes_comites("smt")

#> # A tibble: 106 × 8
#>    data_coleta_dados site_coleta  comite n_ugrhi organizacao_repr… nome    email
#>    <date>            <glue>       <chr>    <dbl> <chr>             <chr>   <chr>
#>  1 2021-07-23        https://sig… Soroc…      10 Prefeitura Munic… Laerte… cole…
#>  2 2021-07-23        https://sig… Soroc…      10 UFSCar - Campus … André … cole…
#>  3 2021-07-23        https://sig… Soroc…      10 DAEE - Departame… Jodhi … cole…
#>  4 2021-07-23        https://sig… Soroc…      10 DAEE - Departame… Arlei … cole…
#>  5 2021-07-23        https://sig… Soroc…      10 CETESB - Companh… Rosâng… NA   
#>  6 2021-07-23        https://sig… Soroc…      10 CETESB - Companh… Rafael… NA   
#>  7 2021-07-23        https://sig… Soroc…      10 Instituto de Pes… José L… NA   
#>  8 2021-07-23        https://sig… Soroc…      10 Instituto de Pes… Gerson… NA   
#>  9 2021-07-23        https://sig… Soroc…      10 DAEE - Departame… Jodhi … NA   
#> 10 2021-07-23        https://sig… Soroc…      10 DAEE - Departame… Arlei … NA   
#> # … with 96 more rows, and 1 more variable: cargo <chr>